requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.0.0
djoser==2.1.0
django-filter==23.1
python-dotenv==1.0.0